cleaned up a bit
git-svn-id: https://scm.mmbase.org/mmbase/trunk@45315 91430c5f-d668-0410-9e7d-dd663366dcfd

+++ core/src/main/java/org/mmbase/core/util/StorageConnector.java
@@ -202,7 +202,7 @@ public class StorageConnector {
         }
         MMObjectNode node = null;
 
-        Integer numberValue = Integer.valueOf(number);
+        Integer numberValue = number;
         // try cache if indicated to do so
         node = builder.getNodeFromCache(numberValue);
         if (node != null) {
@@ -573,7 +573,7 @@ public class StorageConnector {
         while (resultsIterator.hasNext()) {
             MMObjectNode node = resultsIterator.next();
             Integer number = node.getNumber();
-            if(number.intValue() < 0) {
+            if(number < 0) {
                 // never happened to me, and never should!
                 LOG.error("invalid node found, node number was invalid:" + node.getNumber()+", storage invalid?");
                 // dont know what to do with this node,...
