+++ core/src/main/java/org/mmbase/module/corebuilders/OAlias.java
@@ -90,7 +90,7 @@ public class OAlias extends MMObjectBuilder {
                 return -1;
             }
         } else {
-            return nodeNumber.intValue();
+            return nodeNumber;
         }
     }
